+++ src/extensions/rules/rulesext.cpp
@@ -74,7 +74,8 @@ RulesClassExtension::RulesClassExtension(const RulesClass *this_ptr) :
     IsMPPrePlacedConYards(false),
     IsBuildOffAlly(true),
     IsShowSuperWeaponTimers(true),
-    IceStrength(0)
+    IceStrength(0),
+    WeedPipIndex(1)
 {
     //if (this_ptr) EXT_DEBUG_TRACE("RulesClassExtension::RulesClassExtension - 0x%08X\n", (uintptr_t)(ThisPtr));
 
@@ -89,6 +90,13 @@ RulesClassExtension::RulesClassExtension(const RulesClass *this_ptr) :
      */
     This()->EngineerDamage = 1.0f / 3;                    // Amount of damage an engineer does.
     This()->EngineerCaptureLevel = This()->ConditionRed;  // Building damage level before engineer can capture.
+
+    MaxPips = TypeList<int>(5);
+    MaxPips.Add(5);     // PIP_AMMO
+    MaxPips.Add(5);     // PIP_TIBERIUM
+    MaxPips.Add(5);     // PIP_PASSENGERS
+    MaxPips.Add(10);    // PIP_POWER
+    MaxPips.Add(8);     // PIP_CHARGE
 }
 
 
@@ -98,7 +106,8 @@ RulesClassExtension::RulesClassExtension(const RulesClass *this_ptr) :
  *  @author: CCHyper
  */
 RulesClassExtension::RulesClassExtension(const NoInitClass &noinit) :
-    GlobalExtensionClass(noinit)
+    GlobalExtensionClass(noinit),
+    MaxPips(noinit)
 {
     //EXT_DEBUG_TRACE("RulesClassExtension::RulesClassExtension(NoInitClass) - 0x%08X\n", (uintptr_t)(ThisPtr));
 }
@@ -129,6 +138,9 @@ HRESULT RulesClassExtension::Load(IStream *pStm)
         return E_FAIL;
     }
 
+    MaxPips.Clear();
+    MaxPips.Load(pStm);
+
     new (this) RulesClassExtension(NoInitClass());
     
     return hr;
@@ -148,6 +160,8 @@ HRESULT RulesClassExtension::Save(IStream *pStm, BOOL fClearDirty)
     if (FAILED(hr)) {
         return hr;
     }
+
+    MaxPips.Save(pStm);
 
     return hr;
 }
@@ -453,6 +467,11 @@ bool RulesClassExtension::AudioVisual(CCINIClass &ini)
     }
 
     IsShowSuperWeaponTimers = ini.Get_Bool(AUDIOVISUAL, "ShowSuperWeaponTimers", IsShowSuperWeaponTimers);
+    WeedPipIndex = ini.Get_Int(AUDIOVISUAL, "WeedPipIndex", WeedPipIndex);
+    MaxPips = ini.Get_Integers(AUDIOVISUAL, "MaxPips", MaxPips);
+
+    for (int i = 0; i < MaxPips.Count(); i++)
+        DEBUG_INFO("%d", MaxPips[i]);
 
     return true;
 }