+++ src/extensions/rules/rulesext.h
@@ -97,4 +97,14 @@ class RulesClassExtension final : public GlobalExtensionClass<RulesClass>
          *  to break from a shot.
          */
         int IceStrength;
+
+        /**
+         *  Storage pip used for weeds.
+         */
+        int WeedPipIndex;
+
+        /**
+         *  Customizable maximum counts for drawing different pips.
+         */
+        TypeList<int> MaxPips;
 };